validate: separated workflow parsing from validation for yadage environments
closes #487
