+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changes
 =======
 
+Version 0.7.5 (UNRELEASED)
+--------------------------
+- Changes workflow validation to display more granular output.
+- Fixes environment image validation not to test repetitively the same image.
+- Changes workflow parameters validation to warn about misused parameters for each step.
+
 Version 0.7.4 (2021-04-28)
 --------------------------
 - Fixes environment image validation info message where UIDs were switched.
